# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ORCID Link Service",
        "description": "The *ORCID Link Service* provides an API to enable the linking of a KBase\n user account to an ORCID account. This \"link\" consists of a [Link Record](#user-content-header_type_linkrecord) which \n contains a KBase username, ORCID id, ORCID access token, and a few other fields. This link record allows\n KBase to create tools and services which utilize the ORCID api to view or modify\n certain aspects of a users ORCID profile.\n\nOnce connected, *ORCID Link* enables certain integrations, including:\n\n- syncing your KBase profile from your ORCID profile\n- creating and managing KBase public Narratives within your ORCID profile",
        "termsOfService": "https://www.kbase.us/about/terms-and-conditions-v2/",
        "contact": {
            "name": "KBase, Lawrence Berkeley National Laboratory, DOE",
            "url": "https://www.kbase.us/",
            "email": "engage@kbase.us"
        },
        "license": {
            "name": "The MIT License",
            "url": "https://github.com/kbase/kb_sdk/blob/develop/LICENSE.md"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/link": {
            "get": {
                "tags": [
                    "link"
                ],
                "summary": "Get Link",
                "description": "Get ORCID Link\n\nReturn the link for the user associated with the KBase auth token passed in\nthe \"Authorization\" header",
                "operationId": "get_link_link_get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the <a href='#user-content-glossary_term_public-link-record'>Public link record</a> for this user; contains no secrets",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LinkRecordPublic"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Link not available for this user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "link"
                ],
                "summary": "Delete Link",
                "description": "Delete ORCID Link\n\nRemoves the link for the user associated with the KBase auth token passed in the \"Authorization\" header",
                "operationId": "delete_link_link_delete",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted the link"
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Link not available for this user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/link/is_linked": {
            "get": {
                "tags": [
                    "link"
                ],
                "summary": "Is Linked",
                "description": "Get whether Is Linked\n\nDetermine if the user associated with the KBase auth token in the \"Authorization\" header has a\nlink to an ORCID account.",
                "operationId": "is_linked_link_is_linked_get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a boolean indicating whether the user account is linked to ORCID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean",
                                    "title": "Response 200 Is Linked Link Is Linked Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/link/share/{username}": {
            "get": {
                "tags": [
                    "link"
                ],
                "summary": "Link Share",
                "description": "Get whether Is Linked\n\nDetermine if the user associated with the KBase auth token in the \"Authorization\" header has a\nlink to an ORCID account.",
                "operationId": "link_share_link_share__username__get",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The username",
                            "title": "Username"
                        },
                        "description": "The username"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the shared portion of an ORCID Link record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LinkingRecordShared"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string",
                        "maxLength": 25,
                        "minLength": 5,
                        "title": "Code",
                        "description": "A unique code associated with this specific type of error"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5,
                        "title": "Title",
                        "description": "A human-readable title for this error; displayable as an error dialog title"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "A human-readable error message, meant to be displayed to an end user or developer"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ServiceBaseModel"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "title",
                    "message"
                ],
                "title": "ErrorResponse",
                "description": "\n    A generic error object used for all error responses.\n\n    See [the error docs](docs/errors.md) for more information.\n    "
            },
...
            "LinkRecordPublic": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "created_at": {
                        "type": "integer",
                        "title": "Created At"
                    },
                    "expires_at": {
                        "type": "integer",
                        "title": "Expires At"
                    },
                    "orcid_auth": {
                        "$ref": "#/components/schemas/ORCIDAuthPublic"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "created_at",
                    "expires_at",
                    "orcid_auth"
                ],
                "title": "LinkRecordPublic"
            },
            "LinkingRecordShared": {
                "properties": {
                    "orcidId": {
                        "type": "string",
                        "title": "Orcidid"
                    }
                },
                "type": "object",
                "required": [
                    "orcidId"
                ],
                "title": "LinkingRecordShared"
            },
...
            "ORCIDAuthPublic": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "scope": {
                        "type": "string",
                        "title": "Scope"
                    },
                    "expires_in": {
                        "type": "integer",
                        "title": "Expires In"
                    },
                    "orcid": {
                        "type": "string",
                        "title": "Orcid"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "scope",
                    "expires_in",
                    "orcid"
                ],
                "title": "ORCIDAuthPublic"
            },
...
            "ServiceBaseModel": {
                "properties": {},
                "type": "object",
                "title": "ServiceBaseModel"
            },
...
        }
    },
...
}